fix ui css sc_014
