+++ src/modules/chief-police/routes/ChiefPoliceRoutes.jsx
@@ -12,13 +12,12 @@ import SC_013_PreliminaryInvestigation from "@chief-police/pages/SC_013_Prelimin
 import SC_023 from "@chief-police/pages/SC_023_PreliminaryEvidence";
 
 import SC_068_InformationProtectionField from "@chief-police/pages/SC_068_InformationProtectionField";
+import SC_014_AddPatrol from "@chief-police/pages/SC_014_AddPatrol";
 const ChiefPoliceRoutes = () => {
   return (
     <Routes>
       <Route element={<MainLayout />}>
-        <Route path="/" element={<SC_068_InformationProtectionField />} />
-
-        <Route path="/" element={<SC_068_InformationProtectionField />} />
+        <Route path="/" element={<SC_014_AddPatrol />} />
       </Route>
     </Routes>
   );
